split print material integration tests
